Edited Roll Holder
I made it so that the roll holder can't take in dice when it has already rolled 3 times. I also fixed a bug where the order of the dice and their indexes got screwed up if a die was moved out of the holder and moved back in the holder while the dice were moving. Then I changed the top total settings to get the score of the numbered sections and instead get the points of the bonus section.